+++ apps/web/src/Calendar.tsx
@@ -394,7 +394,7 @@ export function Calendar({
             ])) as CalendarConfig
             setCalendarConfig(calendarConfig)
             const titles = await getCachedConfigValue(customization, [...collectionAdminPath, "titles"])
-            setCollectionTitle(titles?.collection)
+            setCollectionTitle(titles?.collection || labels.collection)
             const meta = await getCachedConfigValue(customization, [...collectionAdminPath, "meta"])
             setMeta(meta)
             const originalRangeFilter = filters.find((filter) => filter.type === "range")
@@ -1039,9 +1039,9 @@ export function Calendar({
 
     return (
         <>
-            {!formList && (
+            {!formList && (meta?.title || collectionTitle) && (
                 <Helmet>
-                    <title>{`${meta?.title || collectionTitle || labels.collection} - Calendar`}</title>
+                    <title>{`${meta?.title || collectionTitle || ""} - Calendar`}</title>
                     {meta?.description && <meta name="description" content={meta.description} />}
                 </Helmet>
             )}