+++ packages/system-functions/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @stoker-platform/system-functions
 
+## 0.5.61
+
+### Patch Changes
+
+- Updated dependencies
+    - @stoker-platform/types@0.5.35
+    - @stoker-platform/node-client@0.5.55
+    - @stoker-platform/utils@0.5.47
+
 ## 0.5.60
 
 ### Patch Changes